requirements.txt
Flask==2.3.2
gunicorn==20.1.0
google-api-python-client==2.86.0
google-auth==2.23.0
google-auth-httplib2==0.1.0
google-cloud-storage==2.11.0
